% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_funcs.R
\name{adjust_trackmate_multi_track}
\alias{adjust_trackmate_multi_track}
\title{Summarise across multiple \code{TrackMate} tracks.}
\usage{
adjust_trackmate_multi_track(
  list,
  time_col = "datetime",
  multi_spot_fun = list(max = max)
)
}
\arguments{
\item{list}{A list of \code{TrackMate} data.frames.}

\item{time_col}{A character string that refers to the time column within the data}

\item{multi_splot_fun}{A named list of a function(s) to summarise across \code{time_col} with. Defaults to maximum size.}
}
\value{
A with modified values.
}
\description{
Given a set of IncuCyte data that has been produced via the 2 provided Python scripts, if there are multiple tracks per well, summarise down to one.
}
