% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{assign_colour_vec}
\alias{assign_colour_vec}
\title{Assign Colour Vector for Concentration Plotting}
\usage{
assign_colour_vec(data, concentration_column)
}
\arguments{
\item{data}{A data frame with a \code{condition} column.}

\item{concentration_column}{Character string specifying the column containing
drug concentrations in log10 molar format.}
}
\value{
The input data frame with a new \code{colour_vec} column.
}
\description{
Assign Colour Vector for Concentration Plotting
}
\keyword{internal}
