% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{attempt_fit}
\alias{attempt_fit}
\title{Attempt Fit}
\usage{
attempt_fit(args, data, func)
}
\arguments{
\item{args}{A named list of arguments to pass to \code{minpack.lm::nlsLM()}.}

\item{data}{A data frame with columns \code{x} and \code{y}.}

\item{func}{A function that describes a curve in terms of \code{x}
and individual curve parameters.}
}
\description{
Attempt to fit a nonlinear model using \code{minpack.lm::nlsLM()}.
}
\keyword{internal}
