% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funcs.R
\name{auto_convert_numerics}
\alias{auto_convert_numerics}
\title{Automatically converted character vectors to numerics across a dataframe/tibble.}
\usage{
auto_convert_numerics(.df)
}
\arguments{
\item{.df}{A dataframe or dataframe-like object.}
}
\value{
The original \code{.df} with character columns converted to integer/numeric.
}
\description{
Automatically converted character vectors to numerics across a dataframe/tibble.
}
