% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/double_rate_funcs.R
\name{calc_double_rate_group}
\alias{calc_double_rate_group}
\title{Calculate the double rate for area measurements of spheroids over time.}
\usage{
calc_double_rate_group(
  data,
  x_var = "x",
  y_var = "y",
  shared_group = "group",
  n,
  two_n,
  x_lab = "x",
  y_lab = "y"
)
}
\arguments{
\item{data}{A \code{data.frame} or \code{data.frame} extension (tibble) in long format.}

\item{x_var}{A character string that refers to the the \code{x} value within \code{data}.}

\item{y_var}{A character string that refers to the the \code{y} value within \code{data}.}

\item{shared_group}{A string for the column in \code{data} that refers to individual groups to share fitted parameters.}

\item{n}{A character string in \code{shared_group} that refers to samples which represent samples composed of baseline numbers of cells.}

\item{two_n}{A character string in \code{shared_group} that refers to samples which represent samples composed of two times baseline numbers of cells.}
}
\description{
From a set of matched values that represent area/growth readouts where one group is a known double of another, calculate a doubling time. This function was designed to measure spheroid area from IncuCyte images were cells were initially seeded at 1,000 and 2,000 cells separately. It is recommended that data be generated from the \code{fit_model} function with shared parameters, and then \code{predict_data} used to generate a large number of points to aid in calculations.
}
