% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/double_rate_funcs.R
\name{calc_double_rate_passage}
\alias{calc_double_rate_passage}
\title{Calculate the double rate for known cell numbers over time.}
\usage{
calc_double_rate_passage(start_date, end_date, start_n, end_n)
}
\arguments{
\item{start_date}{The date of the first measurement in format \code{YYYY/MM/DD}.}

\item{end_date}{The date of the second measurement in format \code{YYYY/MM/DD}.}

\item{start_n}{Number of cells at first measurement.}

\item{end_n}{Number of cells at second measurement.}
}
\value{
Doubling in hours.
}
\description{
Calculate the double rate for known cell numbers over time.
}
