% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{conc_data}
\alias{conc_data}
\title{Example Concentration Response Dataset}
\format{
A data frame with concentration-response experimental data
}
\usage{
conc_data
}
\description{
A sample dataset demonstrating concentration-response data structure
for use with incur modeling functions.
}
\seealso{
Other example_data: 
\code{\link{double_data}},
\code{\link{snp_data}}
}
\concept{example_data}
\keyword{datasets}
