% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{double_data}
\alias{double_data}
\title{Example Cell Doubling Dataset}
\format{
A data frame with time-series cell growth data
}
\usage{
double_data
}
\description{
A sample dataset for demonstrating cell doubling rate calculations
and growth analysis functions.
}
\seealso{
Other example_data: 
\code{\link{conc_data}}
}
\concept{example_data}
\keyword{datasets}
