% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{find_identical_to_column}
\alias{find_identical_to_column}
\title{Find Identical Columns}
\usage{
find_identical_to_column(data, target, coerce_to_char = TRUE)
}
\arguments{
\item{data}{A dataframe.}

\item{target}{A character for which to match for.}

\item{coerce_to_char}{TODO: description.}

\item{coerc_to_char}{Logical; convert all columns to characters before match?}
}
\description{
Find identical columns within a dataframe.
}
\keyword{internal}
