% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_funcs.R
\name{find_outlier_indices}
\alias{find_outlier_indices}
\title{Find outliers within a fitted \code{nls} object.}
\usage{
find_outlier_indices(fit, q = 1e-05, .scale_method = "quantile")
}
\arguments{
\item{fit}{A fitted \code{nls} object.}

\item{.scale_method}{The method in which to scale the residuals. The original paper uses the \code{quantile} method. The \code{MAD} implementation is used within the \code{dr4pl} package.}

\item{.q}{The maximum desired FDR.}
}
\value{
A numeric vector of indices individual data points as outliers. This corresponds to the row number of the data.frame used in the original fit.
}
\description{
An implementation of the ROUT method described in Motulsky and Brown (2006).
}
