% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_funcs.R
\name{huber}
\alias{huber}
\title{IRWLS via Huber loss function.}
\usage{
huber(.fit, .arguments, .iter_max = 100, .k = 1.345, .tol = 1e-06)
}
\arguments{
\item{.fit}{A fitted \code{nlsModel} object}

\item{.iter_max}{The number of iterations to perform}

\item{.k}{The tuning constant to be used.}

\item{.tol}{Tolerance level required to be achieved before stopping.}
}
\value{
A fitted \code{nlsModel} object with updated coefficents.
}
\description{
Perform iterative-reweighted least squares regression via Huber loss function.
}
