% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_funcs.R
\name{import_trackmate_spot_data}
\alias{import_trackmate_spot_data}
\title{Import spot data as produced from the \code{merge_incu_channels.py} and \code{trackmate_cellpose_segment.py} scripts}
\usage{
import_trackmate_spot_data(
  path,
  spot_suffix = "-spots.csv",
  time_suffix = "_datetime.csv"
)
}
\arguments{
\item{path}{A directory containing both spot and datetime data.}

\item{spot_suffix}{The suffix used to denote the spot data. This should not need to be modified.}

\item{time_suffix}{The suffix used to denote the datetime data. This should not need to be modified.}
}
\value{
A named list containing spot data per well.
}
\description{
Given a set of IncuCyte data that has been produced via the 2 provided Python scripts, import into R
}
