% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{make_concentration_labels}
\alias{make_concentration_labels}
\title{Generate Concentration Labels for Plots}
\usage{
make_concentration_labels(
  unique_concs,
  negative_control_name,
  negative_control_conc = NA,
  positive_control_name = NA,
  positive_control_conc = NA
)
}
\arguments{
\item{unique_concs}{Numeric vector of unique concentrations.}

\item{negative_control_name}{Character string identifying the negative
control (e.g., "DMSO" or "Vehicle") in \code{treatment_column}.}

\item{negative_control_conc}{Concentration value for negative control.
Default is NA.}

\item{positive_control_name}{Character string identifying the positive
control (e.g., a cytotoxic agent) in \code{treatment_column}. If provided,
NDR values are calculated. Default is NULL.}

\item{positive_control_conc}{Concentration value for positive control.
Default is NA.}
}
\value{
Character vector of formatted labels.
}
\description{
Generate Concentration Labels for Plots
}
\keyword{internal}
