% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_argument_funcs.R
\name{make_formula}
\alias{make_formula}
\title{Make a formula expression from a function}
\usage{
make_formula(func)
}
\arguments{
\item{func}{A function that describes a curve in terms of \code{x} and individual curve parameters.}
}
\value{
A language object in the form of \code{y ~ f()}
}
\description{
From a provided function, create a formula in terms of \verb{y ~ f(\{function arguments\})}.
}
