% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{make_shared_bod}
\alias{make_shared_bod}
\title{Make Shared Body}
\usage{
make_shared_bod(func, group_vec, share_params)
}
\arguments{
\item{func}{A function that describes a curve in terms of \code{x}
and individual curve parameters.}

\item{group_vec}{Character vector of unique group identifiers.}

\item{share_params}{Character vector of parameter names to be shared.}
}
\value{
A call object representing the modified function body.
}
\description{
Generate a function body that implements shared parameter logic using
nested \code{ifelse()} functions.
}
\keyword{internal}
