% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shared_parameter_funcs.R
\name{make_shared_parameters}
\alias{make_shared_parameters}
\title{Run the complete shared parameters pipeline .}
\usage{
make_shared_parameters(
  data,
  shared_params,
  shared_group,
  curve_func,
  start_func
)
}
\arguments{
\item{func}{A function for which arguments are to be modified.}

\item{group}{A character vector that contains all unique grouping values.}

\item{params}{A character vector of arguments in \code{func} that are to be shared across \code{groups}.}
}
\value{
An list data contains in order:
\itemize{
\item Original \code{data} with modified columns to represent groups.
\item A custom \code{curve_func} with modified formals and body.
\item A named list of new starting values for each group.
}
}
\description{
Create a new list of starting values, and custom function that allows for shared parameters across groups.
}
