% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{prep_args}
\alias{prep_args}
\title{Prepare Arguments}
\usage{
prep_args(model_func, start_values, lower_bounds = NA, upper_bounds = NA)
}
\arguments{
\item{model_func}{A function that describes a curve in terms of \code{x}
and individual curve parameters.}

\item{start_values}{A named list of starting parameter values.}

\item{lower_bounds}{Optional named list of lower parameter bounds.}

\item{upper_bounds}{Optional named list of upper parameter bounds.}
}
\value{
A named list to be injected into \code{minpack.lm::nlsLM()} containing:
\describe{
\item{formula}{Model formula object}
\item{start}{Starting parameter values}
\item{lower}{Lower parameter bounds (if specified)}
\item{upper}{Upper parameter bounds (if specified)}
}
}
\description{
Assemble all necessary arguments for \code{minpack.lm::nlsLM()} including formula,
starting values, and optional parameter bounds.
}
\keyword{internal}
