% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{prep_data}
\alias{prep_data}
\title{Prepare Data}
\usage{
prep_data(data, x_var, y_var)
}
\arguments{
\item{data}{A dataframe.}

\item{x_var}{Character string specifying the column name for the independent
variable (typically time).}

\item{y_var}{Character string specifying the column name for the dependent
variable (e.g., cell count, confluence, or a calculated metric).}
}
\description{
Create \code{x} and \code{y} columns from specified variable names.
}
\keyword{internal}
