% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{snp_data}
\alias{snp_data}
\title{Example SNP Dataset}
\format{
A named list with SNP array data
}
\usage{
snp_data
}
\description{
A sample dataset for demonstrating cell line authentication
functions.
}
\seealso{
Other example_data: 
\code{\link{conc_data}},
\code{\link{double_data}}
}
\concept{example_data}
\keyword{datasets}
