% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_funcs.R
\name{try_fit}
\alias{try_fit}
\title{Attempt to fit a curve via \code{minpack.lm}}
\usage{
try_fit(.data, .curve_func, .arguments)
}
\arguments{
\item{.data}{A A \code{data.frame} or \code{data.frame} extension (tibble) in long format with columns \code{x} and \code{y}.}

\item{.curve_func}{A function that describes a curve/model in terms of \code{x}.}

\item{.arguments}{A named list of arguments to be injected into the \code{minpack.lm::nlsLM} function.}
}
\value{
Either a fitted \code{nls} object if expression is evaluated without error, or an invisible object with class \code{try-error}.
}
\description{
Try to fit a curve within a new environment. See help page on \code{rlang::exec} for background on this approach.
}
