% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_funcs.R
\name{calculate_auc}
\alias{calculate_auc}
\title{Calculate AUC for a fitted model.}
\usage{
calculate_auc(fit, lower, upper)
}
\arguments{
\item{fit}{A fitted \code{nlsLM} object.}

\item{lower}{The lower bound of where to calculate AUC.}

\item{upper}{The upper bound of where to calculate AUC.}
}
\value{
A list of class \code{integrate} as produced by \code{stats::integrate}.
}
\description{
A function that finds area under the curve for a fitted \code{nls} object across a lower and upper bound.
}
