% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_gr.R
\name{calculate_gr}
\alias{calculate_gr}
\title{Calculate GR values.}
\usage{
calculate_gr(best_fit_values, group, control_name, cap = TRUE)
}
\arguments{
\item{best_fit_values}{A fitted \code{nlsLM} object.}

\item{lower}{The lower bound of where to calculate AUC.}

\item{upper}{The upper bound of where to calculate AUC.}
}
\value{
A list of class \code{integrate} as produced by \code{stats::integrate}.
}
\description{
Calculate GR values from
}
