% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{incur_models}
\alias{incur_models}
\title{A collection of models with associated functions for both curves and starting value.}
\format{
A named list of length 10:
\describe{
\item{curve_func}{A function that describes a curve in terms of \code{x}}
\item{start_func}{A function that takes in vectors \code{x} and \code{y} of equal length and produces a named list corresponding to parameters in \code{curve_func}}
}
}
\usage{
incur_models
}
\description{
A collection of models with associated functions for both curves and starting value.
}
\keyword{datasets}
