% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funcs.R
\name{index_of_closest_value}
\alias{index_of_closest_value}
\title{Find the index of a vector closest to a target value.}
\usage{
index_of_closest_value(vector, target)
}
\arguments{
\item{vector}{A numeric vector to search within.}

\item{target}{The target value to find the closest value to.}
}
\value{
An index of \code{vector} that contains the closest value to \code{target}
}
\description{
From a vector of \code{numeric}, find which the index of which value is closest to a desired value.
}
