% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_argument_funcs.R
\name{make_bounds}
\alias{make_bounds}
\title{Make a list of upper/lower bounds.}
\usage{
make_bounds(func, lower = NULL, upper = NULL)
}
\arguments{
\item{func}{A function for which to create a list of bounds from.}

\item{lower}{A named list for which arguments in \code{func} to set the lower bounds for.}

\item{upper}{A named list for which arguments in \code{func} to set the upper bounds for.}

\item{lower_bounds}{A named list that contains the lower bounds for specified parameters. All other lower bounds will be set at \code{-Inf}.}

\item{upper_bounds}{A named list that contains the upper bounds for specified parameters. All other upper bounds will be set at \code{Inf}.}
}
\value{
A named list of all arguments as seen in \code{func} with provided bounds set, and Inf/-Inf for all other upper/lower bounds respectively.
}
\description{
From a function and a list of lower and/or upper bounds to be set, create a named list with bounds set for all arguments in \code{func} as provided, with all others set to Inf/-Inf for upper/lower bounds respectively.
}
