% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{make_formula}
\alias{make_formula}
\title{Make Formula}
\usage{
make_formula(func)
}
\arguments{
\item{func}{A function that describes a curve in terms of \code{x}
and individual curve parameters.}
}
\value{
A language object representing the formula \code{y ~ f(parameters)}.
}
\description{
Generate a formula object in the form \code{y ~ f(arguments)} from a function.
}
\keyword{internal}
