% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_shared_body.R
\name{make_shared_body}
\alias{make_shared_body}
\title{Make shared function body across arguments and individual groups.}
\usage{
make_shared_body(.func, .group, .params)
}
\arguments{
\item{.func}{A function for which the body is to be modified.}

\item{.group}{A character vector that contains all unique grouping values.}

\item{.params}{A character vector of arguments in \code{.func} that are to be shared across \code{.groups}.}
}
\value{
An unevaluated \code{call} that contains a vectorised \code{case_when} to modify shared parameters based on \code{.group} values. This can be used to then create a new function
}
\description{
Make shared function body across arguments and individual groups.
}
