% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_shared_formals.R
\name{make_shared_formals}
\alias{make_shared_formals}
\title{Make shared function formals across arguments and individual groups.}
\usage{
make_shared_formals(.func, .group, .params)
}
\arguments{
\item{.func}{A function for which arguments are to be modified.}

\item{.group}{A character vector that contains all unique grouping values.}

\item{.params}{A character vector of arguments in \code{.func} that are to be shared across \code{.groups}.}
}
\value{
An \code{pairlist} object that that can be used to create a new function
}
\description{
Make shared function formals across arguments and individual groups.
}
