% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{make_shared_formals}
\alias{make_shared_formals}
\title{Make Shared Formals}
\usage{
make_shared_formals(func, group_vec, share_params)
}
\arguments{
\item{func}{A function that describes a curve in terms of \code{x}
and individual curve parameters.}

\item{group_vec}{Character vector of unique group identifiers.}

\item{share_params}{Character vector of parameter names to be shared.}
}
\value{
A pairlist object suitable for creating a new function with
group-specific parameters.
}
\description{
Modify function formal arguments to support shared parameters across groups
by creating group-specific parameter names for non-shared parameters.
}
\keyword{internal}
