% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_funcs.R
\name{plot_model}
\alias{plot_model}
\title{Plot a model fitted with \code{incur}.}
\usage{
plot_model(data, x_var, y_var, fit, return_data = FALSE)
}
\arguments{
\item{data}{A \code{data.frame} or \code{data.frame} extension (tibble) in long format.}

\item{x_var}{A string that refers to the the \code{x} value within \code{data}.}

\item{y_var}{A string that refers to the the \code{y} value within \code{data}.}

\item{fit}{The fitted \code{nlsModel} object.}

\item{return_data}{Whether to instead return the original and predicted data to be plot elsewhere.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot a model fitted with \code{incur}.
}
