% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_funcs.R
\name{predict_data}
\alias{predict_data}
\title{Predict data from a \code{nls} object over a range of set \code{x} vakyes}
\usage{
predict_data(fit, lower_x, upper_x, group = NULL, num_points = 1000)
}
\arguments{
\item{fit}{A \code{nls} object}

\item{lower_x}{The lower \code{x} value to predict from.}

\item{upper_x}{The upper \code{x} value to predict from.}

\item{group}{A character vector that contains each unique grouping value used in the initial fit.}

\item{num_points}{How many points to predict values between \code{lower_x} and \code{upper_x}}
}
\value{
Either a fitted \code{nls} object if expression is evaluated without error, or an invisible object with class \code{try-error}
}
\description{
This function can be used to predict a range of values between set points, and further set the precise number of values between said points. This assumes the \code{fit} object takes values in terms of \code{x}.
}
