% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funcs.R
\name{x_at_y_mid}
\alias{x_at_y_mid}
\title{Find the closest x value at the midpoint of y.}
\usage{
x_at_y_mid(.x, .y)
}
\arguments{
\item{.x}{A numeric vector}

\item{.y}{A numeric vector}
}
\value{
The closest value for \code{x} that represents the midpoint of \code{y}.
}
\description{
For two vectors x and y of equal length, this will find the x value where y is at it's mid point.
}
